+++ CHANGELOG.md
@@ -6,6 +6,15 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.5.4
 
+### Fixed
+
+* `SerialFuture` was broken in python 3.8
+
+### Added
+
+* `_ensure_imgsize` can now use threads
+
+
 ## Version 0.5.3 - Released 2020-02-03
 
 ## Version 0.5.2 - Released 2020-01-31